@@ -1,8 +1,8 @@
 # Be sure to restart your server when you modify this file.
 
-MyCitytrain::Application.config.session_store :cookie_store, :key => '_mycitytrain_session', :expire_after => 25.years
+MycitytrainRails::Application.config.session_store :cookie_store, key: '_mycitytrain.rails_session'
 
 # Use the database for sessions instead of the cookie-based default,
 # which shouldn't be used to store highly confidential information
-# (create the session table with "rake db:sessions:create")
-# MyCitytrain::Application.config.session_store :active_record_store
+# (create the session table with "rails generate session_migration")
+# MycitytrainRails::Application.config.session_store :active_record_store
